+++ package.json
@@ -12,8 +12,8 @@
     "load-data": "docker exec -it fastify-postgres psql -c 'CREATE TABLE users(id serial PRIMARY KEY, username VARCHAR (50) NOT NULL);' -U postgres -d postgres",
     "postgres": "docker run -p 5432:5432 --name fastify-postgres -e POSTGRES_PASSWORD=postgres -d postgres:11-alpine",
     "test": "npm run test:unit && npm run test:typescript",
-    "test:unit": "borp --reporter=./test/test-reporter.mjs --coverage --check-coverage --lines 100",
-    "test:report": "standard && c8 --reporter html borp --reporter=./test/test-reporter.mjs",
+    "test:unit": "c8 --100 node --test",
+    "test:report": "standard && c8 --reporter html node --test",
     "test:typescript": "tsd"
   },
   "repository": {
@@ -40,13 +40,11 @@
   "devDependencies": {
     "@tsconfig/node10": "^1.0.9",
     "@types/pg": "^8.11.4",
-    "borp": "^0.18.0",
     "c8": "^10.1.2",
     "fastify": "^5.0.0",
     "pg": "^8.11.3",
     "pg-native": "^3.0.1",
     "standard": "^17.1.0",
-    "tap": "^21.0.1",
     "tsd": "^0.31.1",
     "typescript": "~5.4.3"
   },
